feat: refactor into separate latch proxy
